fix: prevent widget re-rendering on bridge conditions change
fix: prevent widget re-rendering on bridge conditions change
chore: simplify logic for showing or hiding to address field based on bridge conditions
