+++ src/components/Widgets/Widget.tsx
@@ -13,6 +13,7 @@ import {
   HiddenUI,
   LiFiWidget,
   WidgetSkeleton as LifiWidgetSkeleton,
+  RequiredUI,
 } from '@lifi/widget';
 import { PrefetchKind } from 'next/dist/client/components/router-reducer/router-reducer-types';
 import { useRouter } from 'next/navigation';
@@ -23,7 +24,7 @@ import { publicRPCList } from 'src/const/rpcList';
 import { ThemesMap } from 'src/const/themesMap';
 import { useMemelist } from 'src/hooks/useMemelist';
 import { useWelcomeScreen } from 'src/hooks/useWelcomeScreen';
-import { useWidgetSelection } from 'src/hooks/useWidgetSelection';
+import { useBridgeConditions } from 'src/hooks/useBridgeConditions';
 import { useActiveTabStore } from 'src/stores/activeTab';
 import { useContributionStore } from 'src/stores/contribution/ContributionStore';
 import { themeAllowChains, WidgetWrapper } from '.';
@@ -51,9 +52,8 @@ export function Widget({
   ]);
   const formRef = useRef<FormState>(null);
   const wrapperRef = useRef<HTMLDivElement>(null);
-  const { bridgeConditions } = useWidgetSelection({
+  const bridgeConditions = useBridgeConditions({
     formRef,
-    wrapperRef,
     allowToChains,
     configThemeChains: configTheme?.chains,
   });
@@ -180,15 +180,11 @@ export function Widget({
       },
       hiddenUI: [
         ...(configTheme?.hiddenUI ?? []),
-        ...(bridgeConditions.shouldHideToAddress ? [HiddenUI.ToAddress] : []),
         HiddenUI.Appearance,
         HiddenUI.Language,
         HiddenUI.PoweredBy,
         HiddenUI.WalletMenu,
       ],
-      requiredUI: bridgeConditions.shouldRequireToAddress
-        ? ['toAddress']
-        : undefined,
       appearance: widgetTheme.config.appearance,
       theme: widgetTheme.config.theme,
       keyPrefix: `jumper-${starterVariant}`,
@@ -291,9 +287,19 @@ export function Widget({
     widgetTheme.config.appearance,
     widgetTheme.config.theme,
     integratorStringByType,
-    bridgeConditions,
     theme,
   ]);
+
+  if (bridgeConditions.isAGWToNonABSChain) {
+    config.requiredUI = [...(config.requiredUI || []), RequiredUI.ToAddress];
+  }
+
+  if (
+    bridgeConditions.isBridgeFromHypeToArbNativeUSDC ||
+    bridgeConditions.isBridgeFromEvmToHype
+  ) {
+    config.hiddenUI = [...(config.hiddenUI || []), HiddenUI.ToAddress];
+  }
 
   return (
     <WidgetWrapper
